Update Kernel, keep L/M/N and K zImage and modules
ac32525 defconfig: tuna: Modular OMAPLFB
2687bd7 defconfig: espresos: Modular OMAPLFB
24c4392 omaplfb: Prepare for modular build (omaplfb.ko)
d2414c7 SELinux: do not allocate stack space for AVC data unless needed
78c76be selinux: don't transliterate MAY_NOT_BLOCK to IPERM_FLAG_RCU
1a07dd6 lsm_audit: don't specify the audit pre/post callbacks in 'struct common_audit_data'
3e52d03 Yama: add PR_SET_PTRACER_ANY
575c5a9 security: Yama LSM
63ee3e8 capabilities: initialize has_cap
e28294b capabilities: do not grant full privs for setuid w/ file caps + no effective caps
64d62ad selinux: fix off-by-one in setprocattr
288c0a4 inet_diag: Partly rename inet_ to sock_
484951f rcu: treewide: Do not use rcu_read_lock_held when calling rcu_dereference_check
995867f selinux: Fix Kernel panic due to NULL pointer dereference
9c23e17 selinux: remove unused|deprecated includes
364f5cc4 SELinux: check OPEN on truncate calls
9e2bea0 SELinux: include definition of new capabilities
81ef3d4 SELinux: remove avd from selinux_audit_data
2e477cf LSM: shrink the common_audit_data data union
2efa0b5 get rid of pointless includes of ext2_fs.h
393b3e3 selinuxfs: merge dentry allocation into sel_make_dir()
5940c71 SELinux: selinux/xfrm.h needs net/flow.h
42ef646 selinux: Casting (void *) value returned by kmalloc is useless
550e224 selinuxfs: remove custom hex_to_bin()
169c48d selinux: sparse fix: fix several warnings in the security server code
8dde0ae selinux: sparse fix: include selinux.h in exports.c
6c2c392 selinux: sparse fix: fix warnings in netlink code
0100b82 selinux: sparse fix: eliminate warnings for selinuxfs
928f5e0 selinux: sparse fix: declare selinux_disable() in security.h
0da4356 selinux: sparse fix: move selinux_complete_init
5b31a53 selinux: sparse fix: make selinux_secmark_refcount static
a19256d doc: Update the email address for Paul Moore in various source files
2e90e97 security,rcu: Convert call_rcu(sel_netport_free) to kfree_rcu()
9b59fb3 security,rcu: Convert call_rcu(sel_netnode_free) to kfree_rcu()
1f5b1f9 ->permission() sanitizing: don't pass flags to ->inode_permission()
381ff1a audit: treat s_id as an untrusted string

+++ make_anykernel.sh
@@ -2,23 +2,34 @@
 
 DATE=`eval date +%Y`-`eval date +%m`-`eval date +%d`-`eval date +%H`-`eval date +%M`
 
-if [ -f ./tmp/anykernel.sh ]; then
-	cp -fr ./tmp/anykernel.sh ./anykernel.sh
-fi
+if [ -f ./tuna_defconfig-bin/zImage ]; then
 
-if [[ ! -z $1 ]]; then
-	PACKAGENAME=AnyKernel-$DATE-tuna-$1
-else
-	PACKAGENAME=AnyKernel-$DATE-tuna
-fi
+	cp -fr ./tuna_defconfig-bin/zImage ./
+	cp -fr ./tuna_defconfig-bin/modules ./
+	
+	if [ -f ./tmp/anykernel.sh ]; then
+		cp -fr ./tmp/anykernel.sh ./anykernel.sh
+	fi
 
-if [ -f $PACKAGENAME.zip ]; then
+	if [[ ! -z $1 ]]; then
+		PACKAGENAME=AnyKernel-$DATE-tuna-$1
+	else
+		PACKAGENAME=AnyKernel-$DATE-tuna
+	fi
+
+	if [ -f $PACKAGENAME.zip ]; then
 	echo "$PACKAGENAME exist already... skipping."
-else
-	zip -r $PACKAGENAME.zip META-INF/ modules/ tools/ anykernel.sh zImage
-fi
+	else
+		zip -r $PACKAGENAME.zip META-INF/ modules/ tools/ anykernel.sh zImage
+	fi
 
-if [ -f ./anykernel.sh ]; then
-	echo "cleanup"
-	rm -rf ./anykernel.sh
+	if [ -f ./anykernel.sh ]; then
+		echo "cleanup"
+		rm -rf ./anykernel.sh
+		rm -rf modules
+		rm -rf zImage
+	fi
+else
+	echo "zImage does not exits"
+	exit 1
 fi
